requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=9.5.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
stripe>=5.4.0
python-decouple>=3.8
django-extensions>=3.2.0
gunicorn>=20.1.0
whitenoise>=6.4.0
django-filter>=23.2
django-taggit>=4.0.0
django-ckeditor-5>=0.2.18
celery>=5.2.0
redis>=4.5.0
channels>=4.0.0
channels-redis>=4.1.0
reportlab>=4.0.0
qrcode>=7.4.0
geopy>=2.4.1
django-ratelimit>=4.1.0